Add notifications, reviews, talent browse & EO UI
Introduce notifications and reviews support plus EO talent browsing and applicant UX improvements. Added NotificationModel and NotificationRepository (fetch/mark read, mark-all), ReviewRepository (create review), new API endpoints and booking detail helper in api_endpoints.dart. Expanded dashboard providers with notification/review/talent list providers, unread count, and talent list filters. Implemented talent browsing page (EO) and enhanced EO applicants page with detailed applicant modal, accept-with-price dialog, improved accept/reject flow, success/error UX and related repository calls. Misc: hardened EventModel JSON parsing, tolerant RecommendationModel flag parsing, ensured offeredPrice is sent as int, changed app label in AndroidManifest, and added a compile.bat helper script.

+++ lib/features/dashboard/data/invitation_repository.dart
@@ -53,7 +53,7 @@ class InvitationRepository {
       data: {
         'event_id': eventId,
         'talent_id': talentId,
-        'offered_price': offeredPrice,
+        'offered_price': offeredPrice.toInt(),
       },
       parser: (_) {},
     );
